Added method getGearRatio

+++ src/main/java/com/nugumanov/gearcalculator/repos/GearRepo.java
@@ -3,5 +3,6 @@ package com.nugumanov.gearcalculator.repos;
 import com.nugumanov.gearcalculator.domain.Gear;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GearRepo extends CrudRepository<Gear, Long> {
+public interface GearRepo extends CrudRepository<Gear, Integer> {
+
 }
